[hal/driver] restore the modification of 74hc165 to keep it compatible, add rename gpio_i2c
