+++ example/mscboot/project/iar/mscboot.ewp
@@ -6693,10 +6693,10 @@
                         <group>
                             <name>74hc165</name>
                             <file>
-                                <name>$PROJ_DIR$\..\..\..\..\source\hal\driver\vsf\gpio\74hc165\vsf_74hc165_gpio.c</name>
+                                <name>$PROJ_DIR$\..\..\..\..\source\hal\driver\vsf\gpio\74hc165\vsf_gpio_74hc165.c</name>
                             </file>
                             <file>
-                                <name>$PROJ_DIR$\..\..\..\..\source\hal\driver\vsf\gpio\74hc165\vsf_74hc165_gpio.h</name>
+                                <name>$PROJ_DIR$\..\..\..\..\source\hal\driver\vsf\gpio\74hc165\vsf_gpio_74hc165.h</name>
                             </file>
                         </group>
                     </group>
