+++ source/hal/driver/vsf/CMakeLists.txt
@@ -2,3 +2,4 @@
 
 add_subdirectory(distbus)
 add_subdirectory(gpio)
+add_subdirectory(i2c)
