+++ source/hal/driver/vsf/i2c/CMakeLists.txt
@@ -1,0 +1,3 @@
+# CMakeLists head
+
+add_subdirectory(gpio_i2c)
