+++ source/hal/driver/vsf/i2c/gpio_i2c/CMakeLists.txt
@@ -1,0 +1,5 @@
+# CMakeLists head
+
+target_sources(${VSF_LIB_NAME} INTERFACE
+    vsf_gpio_i2c.c
+)
